+++ app/items/[uuid]/page.tsx
@@ -2,27 +2,43 @@ import React from "react";
 import { Metadata } from "next";
 import PageLayout from "../../src/components/pageLayout/pageLayout";
 import Item from "../../src/components/items/item";
-import { RepoApi } from "../../src/utils/apiClients";
+import { RepoApi, CollectionsApi } from "../../src/utils/apiClients";
 import { createAdobeAnalyticsPageName } from "../../src/utils/utils";
 import { ItemModel } from "../../src/models/item";
+import { ItemPage } from "@/src/components/pages/itemPage/itemPage";
 
 type ItemProps = {
   params: {
     uuid: string;
     item: ItemModel;
   };
+  searchParams: { type: string };
 };
 
-const getItemModel = async (uuid: string) => {
+let item;
+
+// const getItemModel = async (uuid: string) => {
+//   const data = await RepoApi.getItemData(uuid);
+//   return new ItemModel(data, uuid);
+//   // return item;
+// };
+
+const getItem = async (uuid: string) => {
   const data = await RepoApi.getItemData(uuid);
-  const item = new ItemModel(data);
-  return item;
+  // const item = new ItemModel(data, uuid);
+  return data;
+};
+
+const getItemManifest = async (uuid: string) => {
+  const data = await CollectionsApi.getManifestForItemUUID(uuid);
+  // const item = new ItemModel(data, uuid);
+  return data;
 };
 
 export async function generateMetadata({
   params,
 }: ItemProps): Promise<Metadata> {
-  const item = await getItemModel(params.uuid);
+  item = await getItem(params.uuid);
   params.item = item;
   return {
     title: `${item.title} - NYPL Digital Collections`, //should be item title
@@ -32,8 +48,11 @@ export async function generateMetadata({
   };
 }
 
-export default async function ItemPage({ params }: ItemProps) {
-  const item = await getItemModel(params.uuid);
+export default async function ItemViewer({ params, searchParams }: ItemProps) {
+  const data = await getItem(params.uuid);
+  const manifest = await getItemManifest(params.uuid);
+  const item = new ItemModel(data, params.uuid, manifest);
+
   return (
     <PageLayout
       activePage="item"
@@ -47,7 +66,12 @@ export default async function ItemPage({ params }: ItemProps) {
       ]}
       adobeAnalyticsPageName={createAdobeAnalyticsPageName("items", item.title)}
     >
-      <Item item={item} />
+      <ItemPage
+        manifest={manifest}
+        data={data}
+        type={searchParams.type}
+        uuid={params.uuid}
+      />
     </PageLayout>
   );
 }