+++ app/src/components/search/search.tsx
@@ -12,12 +12,13 @@ const Search = () => {
   const [keywords, setKeywords] = useState("");
   const [publicDomainOnly, setPublicDomainOnly] = useState(false);
 
+  // TO DO: update this to use DCFL search url
   const handleSubmit = (event) => {
     event.preventDefault();
     const searchUrl =
       `/search/index?` +
-      (publicDomainOnly ? `utf8=✓&filters%5Brights%5D=pd&` : ``) +
-      `keywords=${encodeURIComponent(keywords)}`;
+      `q=${encodeURIComponent(keywords)}` +
+      (publicDomainOnly ? `&filters=%5Brights%3DpublicDomain%5D` : ``);
     router.push(searchUrl);
   };